+++ deno.json
@@ -1,6 +1,6 @@
 {
   "name": "@limette/core",
-  "version": "0.0.15",
+  "version": "0.0.16",
   "exports": {
     ".": "./src/mod.ts",
     "./runtime/is-land.ts": "./src/runtime/is-land.ts",
