+++ init/deno.json
@@ -1,6 +1,6 @@
 {
   "name": "@limette/init",
-  "version": "0.0.15",
+  "version": "0.0.16",
   "exports": {
     ".": "./src/mod.ts"
   },
