+++ src/deps.ts
@@ -9,16 +9,13 @@ export type {
 } from "jsr:@oak/oak@16.1.0";
 
 export { default as esbuild } from "npm:esbuild@0.23.1";
-export { denoPlugins } from "jsr:@luca/esbuild-deno-loader@0.10.3";
+export { denoPlugins } from "jsr:@luca/esbuild-deno-loader@0.11.0";
 export { parseImports } from "npm:parse-imports@2.1.1";
-
-export * as acorn from "npm:acorn@^8.14.0";
-export * as acornWalk from "npm:acorn-walk@8.3.4";
 
 export { DOMParser, Element } from "jsr:@b-fuze/deno-dom@0.1.48";
 
 export { walk } from "jsr:@std/fs@1.0.2/walk";
-export { parse, join } from "jsr:@std/path@1.0.3";
+export { parse, join, isAbsolute, dirname } from "jsr:@std/path@1.0.3";
 export { emptyDir } from "jsr:@std/fs@1.0.2/empty-dir";
 export { ensureFile } from "jsr:@std/fs@1.0.2/ensure-file";
 export { encodeHex } from "jsr:@std/encoding@1.0.3/hex";
